broker_demo/broker_app/outbound_api/requirements.txt
attrs==19.1.0
blinker==1.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==1.1.15
filelock==3.0.10
Flask==1.0.2
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
MarkupSafe==1.1.1
pyrsistent==0.14.11
python-dateutil==2.6.0
psycopg2-binary==2.8.2
requests==2.21.0
sentry-sdk==0.7.7
swagger-spec-validator==2.4.3
swagger-ui-bundle==0.0.3
toml==0.10.0
tox==3.7.0
urllib3==1.24.1
virtualenv==16.4.3
Werkzeug==0.14.1
PyYAML==3.13
SQLAlchemy==1.3.1
alembic==1.0.8
Flask-SQLAlchemy==2.3.2
